requirements.txt
pytorch
langchain
openai
streamlit
streamlit-chat
tiktoken
sentence-transformers
pinecone-client
pinecone-text

